requirements.txt
# FastAPI and Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart

# Data Validation
pydantic>=2.0.0
pydantic-settings

# Database
supabase>=2.0.0

# Redis Cache
redis[hiredis]==5.0.1

# AI & LLM Libraries
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.10
tiktoken

# Embeddings & Vector Search
sentence-transformers
faiss-cpu

# Utilities
python-dotenv
requests
httpx

# Testing
pytest
pytest-asyncio

# Data Processing
pandas
numpy>=1.24.0

# Logging
colorlog

# Additional Dependencies
pytz
dateparsernv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
python-multipart==0.0.6

# Date & Time

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
tenacity>=8.2.0
